Fix validation to accept float instead of int and correctly parse on sanitization

+++ app/api/entities/validation/metadataValidators.js
@@ -65,7 +65,7 @@ const isValidLinkField = value =>
   isString(value.label) && isString(value.url) && ((value.label && value.url) || !value.label);
 
 const validateNumericProperty = value =>
-  isNumber(value) || value === '' || (isString(value) && `${parseInt(value, 10)}` === value);
+  isNumber(value) || value === '' || (isString(value) && `${parseFloat(value, 10)}` === value);
 
 const validateMultiDateProperty = value =>
   Array.isArray(value) && value.every(item => isNumber(item.value) || isNull(item.value));
